feat: new error message

+++ 3. Backend/3. postgresql/src/controllers/weekendproduct.controller.js
@@ -4,12 +4,20 @@ class ProductController {
     // Create a new product
     addProduct = async (req, res) => {
         const { product_name, price } = req.body;
+
+        if (!product_name || !price) {
+            return res.status(400).json({
+                error: `Request Failed: "product_name" or "price" may be missing`
+            });
+        }
+
         try {
             const { rows } = await pool.query(
                 `INSERT INTO "Products" (product_name, price) VALUES ($1, $2) RETURNING *`,
                 [product_name, price]
             );
             res.status(201).json(rows[0]);
+
         } catch (error) {
             res.status(500).json({ error: "Internal Server Error" });
         }
@@ -19,6 +27,11 @@ class ProductController {
     updateProduct = async (req, res) => {
         const { id } = req.params;
         const { product_name, price } = req.body;
+
+        if (!id) return res.status(400).json({error: "product_id is missing"});
+        if (!product_name) return res.status(400).json({error: "product_name is missing"});
+        if (!price) return res.status(400).json({error: "price is missing"});
+
         try {
             const { rowCount } = await pool.query(
                 `UPDATE "Products" SET product_name = $1, price = $2 WHERE product_id = $3`,
@@ -30,6 +43,7 @@ class ProductController {
             }
 
             res.status(200).json({ product_name, price });
+
         } catch (error) {
             res.status(500).json({ error: "Internal Server Error" });
         }
@@ -38,6 +52,9 @@ class ProductController {
     // Delete a product by ID
     deleteProduct = async (req, res) => {
         const { id } = req.params;
+
+        if (!id) return res.status(400).json({error: "product_id is missing"});
+
         try {
             const { rowCount } = await pool.query(
                 `DELETE FROM "Products" WHERE product_id = $1`,
